server/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
pydantic
pydantic-settings
deepgram-sdk<5.0.0
aiohttp
groq
redis
tavily-python
numpy
google-generativeai
python-json-logger
webrtcvad
